+++ src/Kelvin.Server/Integration/Weather/OpenMeteo/OpenMeteoDailyForecast.cs
@@ -10,10 +10,10 @@ public sealed partial class OpenMeteoWeatherApi
     public List<string> Time { get; set; } = [];
 
     [JsonPropertyName("temperature_2m_max")]
-    public List<double> Temperature2mMax { get; set; } = [];
+    public List<float> Temperature2mMax { get; set; } = [];
 
     [JsonPropertyName("temperature_2m_min")]
-    public List<double> Temperature2mMin { get; set; } = [];
+    public List<float> Temperature2mMin { get; set; } = [];
 
     [JsonPropertyName("weather_code")]
     public List<int> WeatherCode { get; set; } = [];
